@@ -15,15 +15,16 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
-        "symfony/config": "^2.3|^3.0",
-        "symfony/translation": "^2.3|^3.0|^4.0",
-        "symfony/yaml": "^2.3|^3.0"
+        "php": ">=7.0.0",
+        "symfony/config": "^2.3|^3.0|^4",
+        "symfony/translation": "^2.3|^3.0|^4",
+        "symfony/yaml": "^2.3|^3.0|^4"
     },
     "require-dev": {
-        "thunderer/shortcode": "~0.5",
+        "thunderer/shortcode": "^0.7",
         "phpspec/phpspec": "^2",
-        "phpunit/phpunit": "^4.5|^5.0"
+        "phpunit/phpunit": "^4.5|^5.0",
+        "friendsofphp/php-cs-fixer": "^2.14"
     },
     "config": {
         "bin-dir": "bin"
@@ -36,10 +37,20 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1-dev"
+            "dev-master": "3.1-dev"
         }
     },
     "suggest": {
         "ext-intl": "Required if you are going to use humanizer with locales different than en_EN"
+    },
+    "scripts": {
+        "cs:php:fix": [
+            "php-cs-fixer fix --using-cache=no"
+        ],
+        "test": [
+            "phpspec run --format=pretty",
+            "phpunit",
+            "php-cs-fixer fix -v --dry-run"
+        ]
     }
 }
