@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\DateTime;
 
 use Coduo\PHPHumanizer\DateTime\Unit\Day;
@@ -50,14 +59,14 @@ final class PreciseDifference
     private function calculate()
     {
         /* @var $units \Coduo\PHPHumanizer\DateTime\Unit[] */
-        $units = array(
+        $units = [
             new Year(),
             new Month(),
             new Day(),
             new Hour(),
             new Minute(),
             new Second(),
-        );
+        ];
 
         $diff = $this->fromDate->diff($this->toDate);
 
