@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\DateTime;
 
 use Symfony\Component\Translation\TranslatorInterface;
@@ -27,13 +36,13 @@ final class PreciseFormatter
      */
     public function formatDifference(PreciseDifference $difference, $locale = 'en')
     {
-        $diff = array();
+        $diff = [];
 
         foreach ($difference->getCompoundResults() as $result) {
             $diff[] = $this->translator->transChoice(
                 'compound.'.$result->getUnit()->getName(),
                 $result->getQuantity(),
-                array('%count%' => $result->getQuantity()),
+                ['%count%' => $result->getQuantity()],
                 'difference',
                 $locale
             );
@@ -41,7 +50,7 @@ final class PreciseFormatter
 
         return $this->translator->trans(
             'compound.'.($difference->isPast() ? 'past' : 'future'),
-            array('%value%' => implode(', ', $diff)),
+            ['%value%' => \implode(', ', $diff)],
             'difference',
             $locale
         );
