@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\Number;
 
 final class RomanNumeral
@@ -8,7 +17,7 @@ final class RomanNumeral
     const MAX_VALUE = 3999;
     const ROMAN_STRING_MATCHER = '/^M{0,4}(CM|CD|D?C{0,3})(XC|XL|L?X{0,3})(IX|IV|V?I{0,3})$/';
 
-    private $map = array(
+    private $map = [
         'M' => 1000,
         'CM' => 900,
         'D' => 500,
@@ -22,7 +31,7 @@ final class RomanNumeral
         'V' => 5,
         'IV' => 4,
         'I' => 1,
-    );
+    ];
 
     /**
      * @param $number
@@ -61,12 +70,12 @@ final class RomanNumeral
      */
     public function fromRoman($string)
     {
-        if (mb_strlen($string) === 0 || 0 === preg_match(self::ROMAN_STRING_MATCHER, $string)) {
+        if (\mb_strlen((string) $string) === 0 || 0 === \preg_match(self::ROMAN_STRING_MATCHER, (string) $string)) {
             throw new \InvalidArgumentException();
         }
 
         $total = 0;
-        $i = mb_strlen($string);
+        $i = \mb_strlen($string);
 
         while ($i > 0) {
             $digit = $this->map[$string{--$i}];
