@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\String;
 
 final class HtmlTruncate implements TruncateInterface
@@ -36,7 +45,7 @@ final class HtmlTruncate implements TruncateInterface
      */
     public function truncate($text, $charactersCount)
     {
-        $strippedText = strip_tags($text, $this->allowedTags);
+        $strippedText = \strip_tags($text, $this->allowedTags);
 
         return $this->truncateHtml($strippedText, $charactersCount);
     }
@@ -56,37 +65,37 @@ final class HtmlTruncate implements TruncateInterface
     {
         $limit = $charactersCount;
         $offset = 0;
-        $tags = array();
+        $tags = [];
 
         // Handle special characters.
-        preg_match_all('/&[a-z]+;/i', strip_tags($string), $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER);
+        \preg_match_all('/&[a-z]+;/i', \strip_tags($string), $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER);
         foreach ($matches as $match) {
             if ($match[0][1] >= $limit) {
                 break;
             }
-            $limit += (mb_strlen($match[0][0]) - 1);
+            $limit += (\mb_strlen($match[0][0]) - 1);
         }
 
         // Handle all the html tags.
-        preg_match_all('/<[^>]+>([^<]*)/', $string, $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER);
+        \preg_match_all('/<[^>]+>([^<]*)/', $string, $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER);
         foreach ($matches as $match) {
             if ($match[0][1] - $offset >= $limit) {
                 break;
             }
 
-            $tag = mb_substr(strtok($match[0][0], " \t\n\r\0\x0B>"), 1);
+            $tag = \mb_substr(\strtok($match[0][0], " \t\n\r\0\x0B>"), 1);
             if ($tag[0] != '/') {
                 $tags[] = $tag;
-            } elseif (end($tags) == mb_substr($tag, 1)) {
-                array_pop($tags);
+            } elseif (\end($tags) == \mb_substr($tag, 1)) {
+                \array_pop($tags);
             }
 
             $offset += $match[1][1] - $match[0][1];
         }
 
-        $newString = mb_substr($string, 0, $limit = min(mb_strlen($string),  $this->breakpoint->calculatePosition($string, $limit + $offset)));
-        $newString .= (mb_strlen($string) > $limit ? $this->append : '');
-        $newString .= (count($tags = array_reverse($tags)) ? '</'.implode('></', $tags).'>' : '');
+        $newString = \mb_substr($string, 0, $limit = \min(\mb_strlen($string), $this->breakpoint->calculatePosition($string, $limit + $offset)));
+        $newString .= (\mb_strlen($string) > $limit ? $this->append : '');
+        $newString .= (\count($tags = \array_reverse($tags)) ? '</'.\implode('></', $tags).'>' : '');
 
         return $newString;
     }
