@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\DateTime;
 
 use Symfony\Component\Translation\TranslatorInterface;
@@ -27,12 +36,12 @@ final class Formatter
      */
     public function formatDifference(Difference $difference, $locale = 'en')
     {
-        $translationKey = sprintf('%s.%s', $difference->getUnit()->getName(), $difference->isPast() ? 'past' : 'future');
+        $translationKey = \sprintf('%s.%s', $difference->getUnit()->getName(), $difference->isPast() ? 'past' : 'future');
 
         return $this->translator->transChoice(
             $translationKey,
             $difference->getQuantity(),
-            array('%count%' => $difference->getQuantity()),
+            ['%count%' => $difference->getQuantity()],
             'difference',
             $locale
         );
