@@ -1,12 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\Resources\Ordinal;
 
 use Coduo\PHPHumanizer\Number\Ordinal\StrategyInterface;
 
 final class EnStrategy implements StrategyInterface
 {
-    /** 
+    /**
      * {@inheritdoc}
      */
     public function isPrefix()
@@ -19,9 +28,9 @@ final class EnStrategy implements StrategyInterface
      */
     public function ordinalIndicator($number)
     {
-        $absNumber = abs((integer) $number);
+        $absNumber = \abs((integer) $number);
 
-        if (in_array(($absNumber % 100), array(11, 12, 13), true)) {
+        if (\in_array(($absNumber % 100), [11, 12, 13], true)) {
             return 'th';
         }
 
