@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\Translator;
 
 use Symfony\Component\Translation\Loader\YamlFileLoader;
@@ -22,7 +31,7 @@ final class Builder
         foreach ($filter as $file) {
             /* @var $file \SplFileInfo */
             $resourceName = $file->getBasename('.yml');
-            list($fileDomain, $fileLocale) = explode('.', $resourceName);
+            list($fileDomain, $fileLocale) = \explode('.', $resourceName);
             $translator->addResource('yml', $file->getPathname(), $fileLocale, $fileDomain);
         }
 
