@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\Tests;
 
 use Coduo\PHPHumanizer\NumberHumanizer;
@@ -281,13 +290,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalIndicatorProvider()
     {
-        return array(
-            array('st', 1),
-            array('nd', 2),
-            array('rd', 23),
-            array('nd', 1002),
-            array('th', -111),
-        );
+        return [
+            ['st', 1],
+            ['nd', 2],
+            ['rd', 23],
+            ['nd', 1002],
+            ['th', -111],
+        ];
     }
 
     /**
@@ -295,13 +304,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalIndicatorDutchProvider()
     {
-        return array(
-            array('e', 1),
-            array('e', 2),
-            array('e', 23),
-            array('e', 1002),
-            array('e', -111),
-        );
+        return [
+            ['e', 1],
+            ['e', 2],
+            ['e', 23],
+            ['e', 1002],
+            ['e', -111],
+        ];
     }
 
     /**
@@ -309,13 +318,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalIndicatorIndonesianProvider()
     {
-        return array(
-            array('ke-', 1),
-            array('ke-', 2),
-            array('ke-', 23),
-            array('ke-', 1002),
-            array('ke-', -111),
-        );
+        return [
+            ['ke-', 1],
+            ['ke-', 2],
+            ['ke-', 23],
+            ['ke-', 1002],
+            ['ke-', -111],
+        ];
     }
 
     /**
@@ -323,13 +332,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataProvider()
     {
-        return array(
-            array('1st', 1),
-            array('2nd', 2),
-            array('23rd', 23),
-            array('1002nd', 1002),
-            array('-111th', -111),
-        );
+        return [
+            ['1st', 1],
+            ['2nd', 2],
+            ['23rd', 23],
+            ['1002nd', 1002],
+            ['-111th', -111],
+        ];
     }
 
     /**
@@ -337,13 +346,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataDutchProvider()
     {
-        return array(
-            array('1e', 1),
-            array('2e', 2),
-            array('23e', 23),
-            array('1002e', 1002),
-            array('-111e', -111),
-        );
+        return [
+            ['1e', 1],
+            ['2e', 2],
+            ['23e', 23],
+            ['1002e', 1002],
+            ['-111e', -111],
+        ];
     }
 
     /**
@@ -351,13 +360,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataIndonesianProvider()
     {
-        return array(
-            array('ke-1', 1),
-            array('ke-2', 2),
-            array('ke-23', 23),
-            array('ke-1002', 1002),
-            array('ke--111', -111),
-        );
+        return [
+            ['ke-1', 1],
+            ['ke-2', 2],
+            ['ke-23', 23],
+            ['ke-1002', 1002],
+            ['ke--111', -111],
+        ];
     }
 
     /**
@@ -365,13 +374,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataPtEsItProvider()
     {
-        return array(
-            array('1o', 1),
-            array('2o', 2),
-            array('23o', 23),
-            array('1002o', 1002),
-            array('-111o', -111),
-        );
+        return [
+            ['1o', 1],
+            ['2o', 2],
+            ['23o', 23],
+            ['1002o', 1002],
+            ['-111o', -111],
+        ];
     }
 
     /**
@@ -379,13 +388,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataGermanProvider()
     {
-        return array(
-            array('1.', 1),
-            array('2.', 2),
-            array('23.', 23),
-            array('1002.', 1002),
-            array('-111.', -111),
-        );
+        return [
+            ['1.', 1],
+            ['2.', 2],
+            ['23.', 23],
+            ['1002.', 1002],
+            ['-111.', -111],
+        ];
     }
 
     /**
@@ -393,26 +402,26 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalizeDataFrenchProvider()
     {
-        return array(
-            array('1er', 1),
-            array('2e', 2),
-            array('23e', 23),
-            array('1002e', 1002),
-            array('-111e', -111),
-        );
+        return [
+            ['1er', 1],
+            ['2e', 2],
+            ['23e', 23],
+            ['1002e', 1002],
+            ['-111e', -111],
+        ];
     }
     /**
      * @return array
      */
     public function ordinalSuffixPtEsItProvider()
     {
-        return array(
-            array('o', 1),
-            array('o', 2),
-            array('o', 23),
-            array('o', 1002),
-            array('o', -111),
-        );
+        return [
+            ['o', 1],
+            ['o', 2],
+            ['o', 23],
+            ['o', 1002],
+            ['o', -111],
+        ];
     }
 
     /**
@@ -420,13 +429,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalSuffixGermanProvider()
     {
-        return array(
-            array('.', 1),
-            array('.', 2),
-            array('.', 23),
-            array('.', 1002),
-            array('.', -111),
-        );
+        return [
+            ['.', 1],
+            ['.', 2],
+            ['.', 23],
+            ['.', 1002],
+            ['.', -111],
+        ];
     }
 
     /**
@@ -434,13 +443,13 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function ordinalSuffixFrenchProvider()
     {
-        return array(
-            array('er', 1),
-            array('e', 2),
-            array('e', 23),
-            array('e', 1002),
-            array('e', -111),
-        );
+        return [
+            ['er', 1],
+            ['e', 2],
+            ['e', 23],
+            ['e', 1002],
+            ['e', -111],
+        ];
     }
 
     /**
@@ -448,22 +457,22 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function binarySuffixDataProvider()
     {
-        return array(
-            array(-1, -1),
-            array("0 bytes", 0),
-            array("1 bytes", 1),
-            array("1 kB", 1024),
-            array("1 kB", 1025),
-            array("1.5 kB", 1536),
-            array("84.4 kB", 86450),
-            array("5 MB", 1048576 * 5),
-            array("2 GB", 1073741824 * 2),
-            array("3 TB", 1099511627776 * 3),
-            array("1.18 PB", 1325899906842624),
+        return [
+            [-1, -1],
+            ['0 bytes', 0],
+            ['1 bytes', 1],
+            ['1 kB', 1024],
+            ['1 kB', 1025],
+            ['1.5 kB', 1536],
+            ['84.4 kB', 86450],
+            ['5 MB', 1048576 * 5],
+            ['2 GB', 1073741824 * 2],
+            ['3 TB', 1099511627776 * 3],
+            ['1.18 PB', 1325899906842624],
 
-            array("1,5 kB", 1536, 'pl'),
-            array("1,18 PB", 1325899906842624, 'pl'),
-        );
+            ['1,5 kB', 1536, 'pl'],
+            ['1,18 PB', 1325899906842624, 'pl'],
+        ];
     }
 
     /**
@@ -471,41 +480,41 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function preciseBinarySuffixDataProvider()
     {
-        return array(
+        return [
             // Negative case
-            array(-1, -1, 3),
+            [-1, -1, 3],
 
             // Byte Cases
-            array("0 bytes", 0, 3),
-            array("1 bytes", 1, 0),
-            array("1023 bytes", 1023, 3),
+            ['0 bytes', 0, 3],
+            ['1 bytes', 1, 0],
+            ['1023 bytes', 1023, 3],
 
             // Kilobyte Cases
-            array('1.000 kB', 1024, 3),
-            array("2 kB", 1588, 0),
-            array("1.6 kB", 1588, 1),
-            array("1.55 kB", 1588, 2),
-            array("1.551 kB", 1588, 3),
+            ['1.000 kB', 1024, 3],
+            ['2 kB', 1588, 0],
+            ['1.6 kB', 1588, 1],
+            ['1.55 kB', 1588, 2],
+            ['1.551 kB', 1588, 3],
 
             // Megabyte Cases
-            array("5.00 MB", (1048576 * 5), 2),
-            array("5.00 MB", (1048576 * 5) + 600, 2),
-            array("5.001 MB", (1048576 * 5) + 600, 3),
+            ['5.00 MB', (1048576 * 5), 2],
+            ['5.00 MB', (1048576 * 5) + 600, 2],
+            ['5.001 MB', (1048576 * 5) + 600, 3],
 
             // Gigabyte Cases
-            array("2 GB", 1073741824 * 2, 0),
-            array("2.0 GB", 1073741824 * 2, 1),
+            ['2 GB', 1073741824 * 2, 0],
+            ['2.0 GB', 1073741824 * 2, 1],
 
             // Terabyte Cases
-            array("3.00 TB", 1099511627776 * 3, 2),
+            ['3.00 TB', 1099511627776 * 3, 2],
 
             // Petabyte Case
-            array("1.178 PB", 1325899906842624, 3),
+            ['1.178 PB', 1325899906842624, 3],
 
             // Locale Cases
-            array("1,500 kB", 1536, 3, 'pl'),
-            array("1,178 PB", 1325899906842624, 3, 'pl'),
-        );
+            ['1,500 kB', 1536, 3, 'pl'],
+            ['1,178 PB', 1325899906842624, 3, 'pl'],
+        ];
     }
 
     /**
@@ -513,20 +522,20 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function metricSuffixDataProvider()
     {
-        return array(
-            array("-1", -1),
-            array("0", 0),
-            array("1", 1),
-            array("101", 101),
-            array("1k", 1000),
-            array("1.2k", 1240),
-            array("1.24M", 1240000),
-            array("3.5M", 3500000),
+        return [
+            ['-1', -1],
+            ['0', 0],
+            ['1', 1],
+            ['101', 101],
+            ['1k', 1000],
+            ['1.2k', 1240],
+            ['1.24M', 1240000],
+            ['3.5M', 3500000],
 
-            array("1,2k", 1240, 'pl'),
-            array("1,24M", 1240000, 'pl'),
-            array("3,5M", 3500000, 'pl'),
-        );
+            ['1,2k', 1240, 'pl'],
+            ['1,24M', 1240000, 'pl'],
+            ['3,5M', 3500000, 'pl'],
+        ];
     }
 
     /**
@@ -534,15 +543,15 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function romanDataProvider()
     {
-        return array(
-            array("I", 1),
-            array("V", 5),
-            array("IX", 9),
-            array("X", 10),
-            array("CXXV", 125),
-            array("MCCC", 1300),
-            array("MMMCMXCIX", 3999),
-        );
+        return [
+            ['I', 1],
+            ['V', 5],
+            ['IX', 9],
+            ['X', 10],
+            ['CXXV', 125],
+            ['MCCC', 1300],
+            ['MMMCMXCIX', 3999],
+        ];
     }
 
     /**
@@ -550,10 +559,10 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function romanExceptionProvider()
     {
-        return array(
-            array(-1),
-            array(4000),
-        );
+        return [
+            [-1],
+            [4000],
+        ];
     }
 
     /**
@@ -561,10 +570,10 @@ class NumberHumanizerTest extends \PHPUnit_Framework_TestCase
      */
     public function arabicExceptionProvider()
     {
-        return array(
-            array(1234),
-            array(""),
-            array("foobar"),
-        );
+        return [
+            [1234],
+            [''],
+            ['foobar'],
+        ];
     }
 }
