@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the PHP Humanizer Library.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Coduo\PHPHumanizer\Tests;
 
 use Coduo\PHPHumanizer\StringHumanizer;
@@ -29,7 +38,7 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
      * @param $charactersCount
      * @param string $append
      */
-    function test_truncate_string_to_word_closest_to_a_certain_number_of_characters($text, $expected, $charactersCount, $append = '')
+    public function test_truncate_string_to_word_closest_to_a_certain_number_of_characters($text, $expected, $charactersCount, $append = '')
     {
         $this->assertEquals($expected, StringHumanizer::truncate($text, $charactersCount, $append));
     }
@@ -43,7 +52,7 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
      * @param $expected
      * @param $append
      */
-    function test_truncate_string_to_word_closest_to_a_certain_number_of_characters_with_html_tags($text, $charactersCount, $allowedTags, $expected, $append = '')
+    public function test_truncate_string_to_word_closest_to_a_certain_number_of_characters_with_html_tags($text, $charactersCount, $allowedTags, $expected, $append = '')
     {
         $this->assertEquals($expected, StringHumanizer::truncateHtml($text, $charactersCount, $allowedTags, $append));
     }
@@ -54,7 +63,7 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
      * @param $text
      * @param $expected
      */
-    function test_remove_all_shortcodes_from_text($text, $expected)
+    public function test_remove_all_shortcodes_from_text($text, $expected)
     {
         $this->assertEquals($expected, StringHumanizer::removeShortcodes($text));
     }
@@ -65,25 +74,25 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
      * @param $text
      * @param $expected
      */
-    function test_remove_only_shortcode_tags_from_text($text, $expected)
+    public function test_remove_only_shortcode_tags_from_text($text, $expected)
     {
         $this->assertEquals($expected, StringHumanizer::removeShortcodeTags($text));
     }
 
     public function removeAllShortcodesProvider()
     {
-        return array(
-            array('some [text] containing [shortcodes /] and [stuff]with[/stuff] content', 'some  containing  and  content'),
-            array('some [text] containing [shortcodes /] and [stuff]with[/stuff] content [/text]', 'some '),
-        );
+        return [
+            ['some [text] containing [shortcodes /] and [stuff]with[/stuff] content', 'some  containing  and  content'],
+            ['some [text] containing [shortcodes /] and [stuff]with[/stuff] content [/text]', 'some '],
+        ];
     }
 
     public function removeShortcodeTagsProvider()
     {
-        return array(
-            array('some [text] containing [shortcodes /] and [stuff]with[/stuff] content', 'some  containing  and with content'),
-            array('some [text] containing [shortcodes /] and [stuff]with[/stuff] content [/text]', 'some  containing  and with content '),
-        );
+        return [
+            ['some [text] containing [shortcodes /] and [stuff]with[/stuff] content', 'some  containing  and with content'],
+            ['some [text] containing [shortcodes /] and [stuff]with[/stuff] content [/text]', 'some  containing  and with content '],
+        ];
     }
 
     /**
@@ -91,15 +100,15 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
      */
     public function humanizeStringProvider()
     {
-        return array(
-            array('news_count', 'News count', true, '_', array('id')),
-            array('user', 'user', false, '_', array('id')),
-            array('news_id', 'News', true, '_', array('id')),
-            array('customer_id', 'Customer id', true, '_', array()),
-            array('news_count', 'News count', true, '_', array('id')),
-            array('news-count', 'News count', true, '-', array('id')),
-            array('news-count', 'news count', false, '-', array('id'))
-        );
+        return [
+            ['news_count', 'News count', true, '_', ['id']],
+            ['user', 'user', false, '_', ['id']],
+            ['news_id', 'News', true, '_', ['id']],
+            ['customer_id', 'Customer id', true, '_', []],
+            ['news_count', 'News count', true, '_', ['id']],
+            ['news-count', 'News count', true, '-', ['id']],
+            ['news-count', 'news count', false, '-', ['id']]
+        ];
     }
 
     /**
@@ -111,24 +120,24 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
         $longText = 'Lorem ipsum dolorem si amet, lorem ipsum. Dolorem sic et nunc.';
         $shortText = 'Short text';
 
-        return array(
-            array($longText, 'Lorem', 2),
-            array($longText, 'Lorem ipsum...', 10, '...'),
-            array($longText, 'Lorem ipsum dolorem si amet, lorem', 30),
-            array($longText, 'Lorem', 0),
-            array($longText, 'Lorem...', 0, '...'),
-            array($longText, 'Lorem ipsum dolorem si amet, lorem ipsum. Dolorem sic et nunc.', -2),
-            array($shortText, "Short...", 1,  '...'),
-            array($shortText, "Short...", 2,  '...'),
-            array($shortText, "Short...", 3,  '...'),
-            array($shortText, "Short...", 4,  '...'),
-            array($shortText, "Short...", 5,  '...'),
-            array($shortText, "Short text", 6,  '...'),
-            array($shortText, "Short text", 7,  '...'),
-            array($shortText, "Short text", 8,  '...'),
-            array($shortText, "Short text", 9,  '...'),
-            array($shortText, "Short text", 10, '...')
-        );
+        return [
+            [$longText, 'Lorem', 2],
+            [$longText, 'Lorem ipsum...', 10, '...'],
+            [$longText, 'Lorem ipsum dolorem si amet, lorem', 30],
+            [$longText, 'Lorem', 0],
+            [$longText, 'Lorem...', 0, '...'],
+            [$longText, 'Lorem ipsum dolorem si amet, lorem ipsum. Dolorem sic et nunc.', -2],
+            [$shortText, 'Short...', 1,  '...'],
+            [$shortText, 'Short...', 2,  '...'],
+            [$shortText, 'Short...', 3,  '...'],
+            [$shortText, 'Short...', 4,  '...'],
+            [$shortText, 'Short...', 5,  '...'],
+            [$shortText, 'Short text', 6,  '...'],
+            [$shortText, 'Short text', 7,  '...'],
+            [$shortText, 'Short text', 8,  '...'],
+            [$shortText, 'Short text', 9,  '...'],
+            [$shortText, 'Short text', 10, '...']
+        ];
     }
 
     /**
@@ -139,20 +148,20 @@ class StringHumanizerTest extends PHPUnit_Framework_TestCase
     {
         $text = '<p><b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup language</a> used to create <a href="/wiki/Web_page" title="Web page">web pages</a>.<sup id="cite_ref-1" class="reference"><a href="#cite_note-1"><span>[</span>1<span>]</span></a></sup> <a href="/wiki/Web_browser" title="Web browser">Web browsers</a> can read HTML files and render them into visible or audible web pages. HTML describes the structure of a <a href="/wiki/Website" title="Website">website</a> <a href="/wiki/Semantic" title="Semantic" class="mw-redirect">semantically</a> along with cues for presentation, making it a markup language, rather than a <a href="/wiki/Programming_language" title="Programming language">programming language</a>.</p>';
 
-        return array(
-            array($text, 3,  '<b><i><u><em><strong><a><span>',  "<b>HyperText</b>"),
-            array($text, 12, '<b><i><u><em><strong><a><span>', "<b>HyperText Markup</b>"),
-            array($text, 30, '<b><i><u><em><strong><a><span>', "<b>HyperText Markup Language</b>, commonly"),
-            array($text, 50, '<b><i><u><em><strong><a><span>', "<b>HyperText Markup Language</b>, commonly referred to as"),
-            array($text, 75, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup</a>'),
-            array($text, 100,'<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup language</a> used to create'),
-            array($text, 3  , '', "HyperText"),
-            array($text, 12 , '', "HyperText Markup"),
-            array($text, 50 , '', "HyperText Markup Language, commonly referred to as"),
-            array($text, 75 , '', "HyperText Markup Language, commonly referred to as HTML, is the standard markup"),
-            array($text, 100, '', "HyperText Markup Language, commonly referred to as HTML, is the standard markup language used to create"),
-            array($text, 50, '', "HyperText Markup Language, commonly referred to as...", '...'),
-            array($text, 75, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup...</a>', '...')
-        );
+        return [
+            [$text, 3,  '<b><i><u><em><strong><a><span>',  '<b>HyperText</b>'],
+            [$text, 12, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup</b>'],
+            [$text, 30, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly'],
+            [$text, 50, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as'],
+            [$text, 75, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup</a>'],
+            [$text, 100,'<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup language</a> used to create'],
+            [$text, 3  , '', 'HyperText'],
+            [$text, 12 , '', 'HyperText Markup'],
+            [$text, 50 , '', 'HyperText Markup Language, commonly referred to as'],
+            [$text, 75 , '', 'HyperText Markup Language, commonly referred to as HTML, is the standard markup'],
+            [$text, 100, '', 'HyperText Markup Language, commonly referred to as HTML, is the standard markup language used to create'],
+            [$text, 50, '', 'HyperText Markup Language, commonly referred to as...', '...'],
+            [$text, 75, '<b><i><u><em><strong><a><span>', '<b>HyperText Markup Language</b>, commonly referred to as <b>HTML</b>, is the standard <a href="/wiki/Markup_language" title="Markup language">markup...</a>', '...']
+        ];
     }
 }
